Lens wide support added, bugs fixed for dynamical propagation

+++ Assets/BACKEND/PropagationSystem/PropagationSystem.cs
@@ -26,12 +26,17 @@ public class PropagationSystem : MonoBehaviour
             {
                 throw new ArgumentOutOfRangeException("The propagation beams are more than the actual. They are given to be : " + value + "Actual size is : " + _discreteBeams.Count);
             }
+            if(value <0)
+            {
+                throw new ArgumentOutOfRangeException("the beams in propagation are beckoming negative in numbers.");
+            }
             __numberOfBeamsInPropagation = value;
         } 
     }
 
     private List<GameObject> _keysForClear = new List<GameObject>();
     private Dictionary<GameObject, Tuple<List<int>, List<RaycastHit>>> _separateObjectInteractions = new Dictionary<GameObject, Tuple<List<int>, List<RaycastHit>>>();
+    private Dictionary<GameObject, bool> _sendToMultiBeamInteractors = new Dictionary<GameObject, bool>();
     private Dictionary<int, Tuple<List<float3>, List<float3>>> _collisionPositionsPerBeam = new Dictionary<int, Tuple<List<float3>, List<float3>>>();
     private Dictionary<GameObject, Tuple<Vector3, Quaternion>> _ObjectOfInteractionPrevPosition = new Dictionary<GameObject, Tuple<Vector3, Quaternion>>();
     private bool _remodelling = false;
@@ -46,7 +51,7 @@ public class PropagationSystem : MonoBehaviour
             int beamPropagationSectionInteger;
             int discreteBeaminteger;
 
-            _remodelling = false;
+            
             ;
             if (CheckForSceneUpdates(out discreteBeaminteger, out beamPropagationSectionInteger))
             {
@@ -55,15 +60,17 @@ public class PropagationSystem : MonoBehaviour
             }
             else
             {
-                if (_separateObjectInteractions.Count != 0)
+                ;
+                if (!_remodelling)
                 {
                     SendMultiBeams();
                 }
+                
                 return;
             }
-            
+            _remodelling = false;
         }
-
+        ;
         var descreteBeamsTointeractIndeces = new List<int>();
 
         for (int i = 0; i < _discreteBeams.Count; i++)
@@ -169,13 +176,27 @@ public class PropagationSystem : MonoBehaviour
             {
                 for (int i = 0; i < _separateObjectInteractions[el].Item1.Count; i++)
                 {
-                    _NumberOfBeamsInPropagation--;
-                    limitedBeamFinalPositionList.Add(_separateObjectInteractions[el].Item2[i].point);
-                    limitedBeamIndecesList.Add(_separateObjectInteractions[el].Item1[i]);
+                    if (!ChechFloat3IsZero( _discreteBeams[_separateObjectInteractions[el].Item1[i]].DirectionOfPropagation) )
+                    {
+                        _NumberOfBeamsInPropagation--;
+                        limitedBeamFinalPositionList.Add(_separateObjectInteractions[el].Item2[i].point);
+                        limitedBeamIndecesList.Add(_separateObjectInteractions[el].Item1[i]);
+                    }
                 }
                 if (!el.TryGetComponent<MultiBeamInteractionI>(out multiBeamInteraction))
                 {
                     _keysForClear.Add(el);
+                }
+                else
+                {
+                    if (!_sendToMultiBeamInteractors.ContainsKey(el))
+                    {
+                        _sendToMultiBeamInteractors.Add(el, false);
+                    }
+                    else
+                    {
+                        _sendToMultiBeamInteractors[el] = false;
+                    }
                 }
                 ;
             }
@@ -287,12 +308,26 @@ public class PropagationSystem : MonoBehaviour
             {
                 for (int i = 0; i < _collisionPositionsPerBeam[el].Item1.Count; i++)
                 {
+                    if(i>= _collisionPositionsPerBeam[el].Item2.Count)
+                    {
+                        ;
+                        if(_collisionPositionsPerBeam[el].Item1.Count - _collisionPositionsPerBeam[el].Item2.Count == 1)
+                        {
+                            break;
+                        }
+                        else
+                        {
+                            throw new ArgumentOutOfRangeException("The difference starts and stops of propagation exceed the expected boundries.");
+                        }
+                    }
                     ray.origin = _collisionPositionsPerBeam[el].Item1[i];
                     ray.direction = _collisionPositionsPerBeam[el].Item2[i] - _collisionPositionsPerBeam[el].Item1[i];
+                     
                     if (Physics.Raycast(ray, out hit, Mathf.Infinity, layerMask))
                     {
                         if (!_ObjectOfInteractionPrevPosition.ContainsKey(hit.collider.gameObject))
                         {
+                            ;
                             beamPropagationSectionInteger = i;
                             discreteBeamInteger = el;
 
@@ -303,6 +338,7 @@ public class PropagationSystem : MonoBehaviour
                         if (hit.collider.gameObject.transform.position != _ObjectOfInteractionPrevPosition[hit.collider.gameObject].Item1 |
                             hit.collider.gameObject.transform.rotation != _ObjectOfInteractionPrevPosition[hit.collider.gameObject].Item2)
                         {
+                            ;
                             beamPropagationSectionInteger = i;
                             discreteBeamInteger = el;
 
@@ -312,6 +348,7 @@ public class PropagationSystem : MonoBehaviour
                         }
                         if (!ChechTwofloat3sEqual(_collisionPositionsPerBeam[el].Item2[i], Vector3ToFloat3(hit.point)))
                         {
+                            ;
                             beamPropagationSectionInteger = i;
                             discreteBeamInteger = el;
                             return true;
@@ -319,13 +356,15 @@ public class PropagationSystem : MonoBehaviour
                     }
                     else
                     {
-                        if(Magniitude(_collisionPositionsPerBeam[el].Item2[i] - _collisionPositionsPerBeam[el].Item1[i]) != maxLengthOfPropagation)
+                        if (Magniitude(_collisionPositionsPerBeam[el].Item2[i] - _collisionPositionsPerBeam[el].Item1[i]) != maxLengthOfPropagation)
                         {
+                            ;
                             beamPropagationSectionInteger = i;
                             discreteBeamInteger = el;
                             return true;
                         }
                     }
+                    
                 }
             }
         }
@@ -379,22 +418,22 @@ public class PropagationSystem : MonoBehaviour
         MultiBeamInteractionI multiBeamInteraction;
         foreach (var el in _separateObjectInteractions.Keys)
         {
-            var discreteBeamsForInteraction = new List<DiscreteBeam>();
-            foreach (var index in _separateObjectInteractions[el].Item1)
+            if (!_sendToMultiBeamInteractors[el])
             {
-                discreteBeamsForInteraction.Add(_discreteBeams[index]);
+                var discreteBeamsForInteraction = new List<DiscreteBeam>();
+                foreach (var index in _separateObjectInteractions[el].Item1)
+                {
+                    discreteBeamsForInteraction.Add(_discreteBeams[index]);
+                }
+
+                if (el.TryGetComponent<MultiBeamInteractionI>(out multiBeamInteraction))
+                {
+                    multiBeamInteraction.MultiBeamGetter(discreteBeamsForInteraction, _separateObjectInteractions[el].Item2);
+                }
+
+                _sendToMultiBeamInteractors[el] = true;
             }
-
-            if (el.TryGetComponent<MultiBeamInteractionI>(out multiBeamInteraction))
-            {
-                multiBeamInteraction.MultiBeamGetter(discreteBeamsForInteraction, _separateObjectInteractions[el].Item2);
-            }
-
-            _keysForClear.Add(el);
-
         }
-
-        ClearInteractionObjectKeys();
     }
 
     private void Project()
@@ -434,6 +473,20 @@ public class PropagationSystem : MonoBehaviour
         }
     }
 
+    private void ClearBeamsFromMultiBeamInteractions(int beamIndex)
+    {
+        foreach(var el in _separateObjectInteractions.Keys)
+        {
+            if (_separateObjectInteractions[el].Item1.Contains(beamIndex))
+            {
+                _sendToMultiBeamInteractors[el] = false;
+                var index = _separateObjectInteractions[el].Item1.IndexOf(beamIndex);
+                _separateObjectInteractions[el].Item1.RemoveAt(index);
+                _separateObjectInteractions[el].Item2.RemoveAt(index);
+            }
+        }
+    }
+
     private void CutBeams(int firstBeamInteger, int beamPropagationSectionInteger)
     {
         Guid parentId = _discreteBeams[firstBeamInteger].ParentId;
@@ -448,59 +501,44 @@ public class PropagationSystem : MonoBehaviour
         {
             if(_discreteBeams[i].ParentId == parentId)
             {
-                ray.origin = _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger];
-                ray.direction = _collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger] - _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger];
-                if (Physics.Raycast(ray, out hit, Mathf.Infinity, layerMask))
+                if (_collisionPositionsPerBeam[i].Item1.Count > beamPropagationSectionInteger)
                 {
-                    if (!ChechTwofloat3sEqual(_collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger], Vector3ToFloat3(hit.point)))
+                    ray.origin = _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger];
+                    ray.direction = _collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger] - _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger];
+                    if (Physics.Raycast(ray, out hit, Mathf.Infinity, layerMask))
                     {
-                        if (ChechFloat3IsZero(_discreteBeams[i].DirectionOfPropagation))
+                        if (!ChechTwofloat3sEqual(_collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger], Vector3ToFloat3(hit.point)))
                         {
-                            _NumberOfBeamsInPropagation++;
+                            if (ChechFloat3IsZero(_discreteBeams[i].DirectionOfPropagation))
+                            {
+                                _NumberOfBeamsInPropagation++;
+                            }
+                            ClearBeamsFromMultiBeamInteractions(i);
+                            _collisionPositionsPerBeam[i].Item1.RemoveRange(beamPropagationSectionInteger + 1, _collisionPositionsPerBeam[i].Item1.Count - 1 - beamPropagationSectionInteger);
+                            _collisionPositionsPerBeam[i].Item2.RemoveRange(beamPropagationSectionInteger, _collisionPositionsPerBeam[i].Item2.Count - beamPropagationSectionInteger);
+                            CutBeam(i, _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger]);
+                            indecesToClear.Add(i);
                         }
-                        _collisionPositionsPerBeam[i].Item1.RemoveRange(beamPropagationSectionInteger+1, _collisionPositionsPerBeam[i].Item1.Count - 1 - beamPropagationSectionInteger);
-                        _collisionPositionsPerBeam[i].Item2.RemoveRange(beamPropagationSectionInteger, _collisionPositionsPerBeam[i].Item2.Count - beamPropagationSectionInteger);
-                        CutBeam(i, _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger]);
-                        indecesToClear.Add(i);
                     }
-                }
-                else
-                {
-                    if(Magniitude(_collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger] - _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger])!=maxLengthOfPropagation)
+                    else
                     {
-                        if (ChechFloat3IsZero(_discreteBeams[i].DirectionOfPropagation))
+                        if (Magniitude(_collisionPositionsPerBeam[i].Item2[beamPropagationSectionInteger] - _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger]) != maxLengthOfPropagation)
                         {
-                            _NumberOfBeamsInPropagation++;
+                            if (ChechFloat3IsZero(_discreteBeams[i].DirectionOfPropagation))
+                            {
+                                _NumberOfBeamsInPropagation++;
+                            }
+                            ;
+                            ClearBeamsFromMultiBeamInteractions(i);
+                            _collisionPositionsPerBeam[i].Item1.RemoveRange(beamPropagationSectionInteger + 1, _collisionPositionsPerBeam[i].Item1.Count - 1 - beamPropagationSectionInteger);
+                            _collisionPositionsPerBeam[i].Item2.RemoveRange(beamPropagationSectionInteger, _collisionPositionsPerBeam[i].Item2.Count - beamPropagationSectionInteger);
+                            CutBeam(i, _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger]);
+                            indecesToClear.Add(i);
                         }
-                        ;
-                        _collisionPositionsPerBeam[i].Item1.RemoveRange(beamPropagationSectionInteger+1, _collisionPositionsPerBeam[i].Item1.Count - 1 - beamPropagationSectionInteger);
-                        _collisionPositionsPerBeam[i].Item2.RemoveRange(beamPropagationSectionInteger, _collisionPositionsPerBeam[i].Item2.Count - beamPropagationSectionInteger);
-                        CutBeam(i, _collisionPositionsPerBeam[i].Item1[beamPropagationSectionInteger]);
-                        indecesToClear.Add(i);
                     }
                 }
             }
         }
-
-        foreach(var el in _separateObjectInteractions.Keys)
-        {
-            MultiBeamInteractionI multiBeamInteraction;
-            if (el.TryGetComponent<MultiBeamInteractionI>(out multiBeamInteraction))
-            {
-                foreach(var indexToClear in indecesToClear)
-                {
-                    var index = _separateObjectInteractions[el].Item1.IndexOf(indexToClear);
-                    if (index != -1)
-                    {
-                        _separateObjectInteractions[el].Item1.RemoveAt(index);
-                        _separateObjectInteractions[el].Item2.RemoveAt(index);
-                    }
-                }
-                
-            }
-                
-        }
-
     }
 
     private void CutBeam(int beamIndex, float3 cutOffPosition)
@@ -531,8 +569,6 @@ public class PropagationSystem : MonoBehaviour
     {
         
     }
-
-    
 
 }
